Column resize fixed
I wanted to avoid that the user is able to resize the column of the
table view. by setting a size
and make the tableview column autoresizing style of auto for the first
column. that means that the first and only column of our tableview is
no longer resizable and fits the tableview.