Settings / Preference improvement
Better binding for the settings, we unbind whenever the window is
closed an delete the instance. this commit allow to create a new
instance of the Settings whenever we open the Window

Settings Model is not need due the the binding

A FormatTransformer is introduced to Bind transformation of NSArray to
NSData and vice-versa.